third_party/WebKit/LayoutTests/virtual/threaded/fast/scroll-behavior/smooth-scroll/main-thread-scrolling-reason-correctness-expected.txt

This test verifies that the ScrollAnimator can schedule animations on the compositor when it adds a temporary main thread scrolling reason.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.getScrollAnimationState(document) == RUNNING_ON_COMPOSITOR is true
PASS internals.mainThreadScrollingReasons(document) == ANIMATING_TEXT is true
PASS internals.getScrollAnimationState(document) == RUNNING_ON_COMPOSITOR_BUT_NEEDS_UPDATE is true
PASS internals.mainThreadScrollingReasons(document) == ANIMATING_TEXT is true
PASS needsUpdateOrRunningOnCompositor(document) is true
PASS document.scrollingElement.scrollTop == 80 became true
PASS internals.mainThreadScrollingReasons(document) == '' is true
PASS successfullyParsed is true

TEST COMPLETE
